rename and move datasource
(cherry picked from commit 128b9ddc135c3cf48c1d4a507d795259a3c7a4de)
