+++ self_description_service.datasource/src/main/java/org/eclipse/slm/self_description_service/datasource/template/TemplateManager.java
@@ -1,4 +1,4 @@
-package org.eclipse.slm.self_description_service.factories.template;
+package org.eclipse.slm.self_description_service.datasource.template;
 
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
