+++ self_description_service.datasource/src/test/resources/templates/Template.aasx
